index: diff: handle missing hash in hash_only mode
Previously if one index had a file without hash and second had no file at all
diff would say that this file was added/deleted in `hash_only=True` mode,
which is incorrect.

See https://github.com/iterative/dvc/pull/8878#discussion_r1086346424
