+++ tests/test_index.py
@@ -10,7 +10,6 @@ from dvc_data.index import (
     add,
     build,
     checkout,
-    diff,
     md5,
     read_db,
     read_json,
@@ -395,63 +394,6 @@ def test_view_traverse(odb):
         ("dir", "subdir"),
         ("dir", "subdir", "bar"),
     ]
-
-
-def test_diff():
-    from dvc_data.index.diff import ADD, DELETE, RENAME, UNCHANGED, Change
-
-    old_foo_key = ("foo",)
-    old_foo_entry = DataIndexEntry(
-        key=old_foo_key,
-        hash_info=HashInfo(
-            name="md5", value="d3b07384d113edec49eaa6238ad5ff00"
-        ),
-    )
-    old_bar_key = ("dir", "subdir", "bar")
-    old_bar_entry = DataIndexEntry(
-        key=old_bar_key,
-        hash_info=HashInfo(
-            name="md5",
-            value="1f69c66028c35037e8bf67e5bc4ceb6a.dir",
-        ),
-    )
-    old = DataIndex({old_foo_key: old_foo_entry, old_bar_key: old_bar_entry})
-
-    assert set(diff(old, old, with_unchanged=True)) == {
-        Change(UNCHANGED, old_foo_entry, old_foo_entry),
-        Change(UNCHANGED, old_bar_entry, old_bar_entry),
-    }
-    assert set(diff(old, old, with_renames=True, with_unchanged=True)) == {
-        Change(UNCHANGED, old_foo_entry, old_foo_entry),
-        Change(UNCHANGED, old_bar_entry, old_bar_entry),
-    }
-
-    new_foo_key = ("data", "FOO")
-    new_foo_entry = DataIndexEntry(
-        key=new_foo_key,
-        hash_info=HashInfo(
-            name="md5", value="d3b07384d113edec49eaa6238ad5ff00"
-        ),
-    )
-    new = DataIndex(
-        {
-            (
-                "data",
-                "FOO",
-            ): new_foo_entry,
-            old_bar_key: old_bar_entry,
-        }
-    )
-
-    assert set(diff(old, new, with_unchanged=True)) == {
-        Change(ADD, None, new_foo_entry),
-        Change(DELETE, old_foo_entry, None),
-        Change(UNCHANGED, old_bar_entry, old_bar_entry),
-    }
-    assert set(diff(old, new, with_renames=True, with_unchanged=True)) == {
-        Change(RENAME, old_foo_entry, new_foo_entry),
-        Change(UNCHANGED, old_bar_entry, old_bar_entry),
-    }
 
 
 def test_update(tmp_upath, odb, as_filesystem):
